web: add LabWAS pivot table on the lab test pages
This adds a new table at the bottom of the lab test pages. It exposes
the same statistics as the LabWAS table on the endpoint pages. The
difference is that this new pivoted LabWAS table has endpoints for
row-index.

Dev notes:
Had to do some refactoring to have the LabWAS table used on both
the lab test pages and the endpoint pages.
Not totally happy with the refactoring, because:
- The lab test / endpoint specializations are a bit scattered around
  (though likely only in 1 file).
- I don't think I understand fully when/how to use Phoenix.Component vs
  LiveView vs LiveComponent.
  For example: Should the endpoint pages and lab test pages be
  LiveView:s?


GitHub: closes #306