web: fix linking to non-existing lab test pages
Before, the search box would do look ups against all the OMOP concepts
in the database. However, we only have stat pages for the "child" OMOP
concept in a OMOP ConceptRelationship. The user was sometimes presented
with "parent" OMOP concepts, for which we don't have any stat page.

Now, the search box does look up only against children of the OMOP
ConceptRelationship, which reflects the OMOP concepts for which we have
data.

GitHub: fix #331